test_data/requirements.txt
langchain==0.3.9
python-dotenv==1.0.1
websockets==14.1
langchain==0.3.10
langchain_community==0.3.9
langchain_openai==0.2.11